fix: remove circular-progress tv to progress
