+++ packages/core/theme/src/components/index.ts
@@ -16,7 +16,6 @@ export * from "./pagination";
 export * from "./toggle";
 export * from "./accordion";
 export * from "./progress";
-export * from "./circular-progress";
 export * from "./input";
 export * from "./dropdown";
 export * from "./image";
